+++ lib/eve_online/esi/universe_bloodlines.rb
@@ -2,7 +2,7 @@
 
 module EveOnline
   module ESI
-    class Bloodlines < Base
+    class UniverseBloodlines < Base
       API_ENDPOINT = 'https://esi.evetech.net/v1/universe/bloodlines/?datasource=%<datasource>s'
 
       def bloodlines